refactor: 폴더 구조 변경 및 기타 리팩토링
