+++ common/src/main/java/kernel/maidlab/common/dto/ResponseDto.java
@@ -2,6 +2,8 @@ package kernel.maidlab.common.dto;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import kernel.maidlab.common.enums.ResponseType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.ResponseEntity;
