+++ common/src/main/java/kernel/maidlab/common/enums/ResponseType.java
@@ -1,4 +1,4 @@
-package kernel.maidlab.common.dto;
+package kernel.maidlab.common.enums;
 
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
