+++ reservation/src/main/java/kernel/maidlab/reservation/entity/Reservation.java
@@ -1,6 +1,5 @@
 package kernel.maidlab.reservation.entity;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import jakarta.persistence.Column;
