+++ reservation/src/main/java/kernel/maidlab/reservation/entity/ServiceDetailType.java
@@ -6,7 +6,7 @@ import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
-import kernel.maidlab.reservation.enums.ServiceType;
+import kernel.maidlab.reservation.entity.enums.ServiceType;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
