+++ reservation/src/main/java/kernel/maidlab/reservation/repository/ServiceDetailTypeRepository.java
@@ -5,7 +5,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import kernel.maidlab.reservation.entity.ServiceDetailType;
-import kernel.maidlab.reservation.enums.ServiceType;
+import kernel.maidlab.reservation.entity.enums.ServiceType;
 
 public interface ServiceDetailTypeRepository extends JpaRepository<ServiceDetailType, Long> {
 	List<ServiceDetailType> findByServiceType(ServiceType type);
